+++ lib/orchestration/summarizers/peer-context.ts
@@ -1,4 +1,4 @@
-import type { AgentTurnSummary } from '../director-prompt';
+import type { AgentTurnSummary } from '../types';
 
 // ==================== Peer Context ====================
